Added option to deactivate point filter
Points to be drawn are automatically filtered using the QuadTree. For
drawing curves, simple filtering by points does not work, hence the
option to deactivate this filtering. This is certainly not an elegant
solution, but it is probably still ok for many situations. Need to
introduce bounds checking based on geometry bounding boxes or similar
(again a simple check won’t work for some curves).