Replaced curve styling with style function
Instead of providing individual stroke style and line width for each
curve, it is now possible to provide a style function that returns a
dictionary with the stroke style and line width. The style function
receives the data for the corresponding the line, enabling dynamic
rendering of the curve style.